+++ source/common/http/http1/codec_impl.cc
@@ -74,7 +74,10 @@ void StreamEncoderImpl::encodeHeaders(const HeaderMap& headers, bool end_stream)
   // Assume we are chunk encoding unless we are passed a content length or this is a header only
   // response. Upper layers generally should strip transfer-encoding since it only applies to
   // HTTP/1.1. The codec will infer it based on the type of response.
-  if (saw_content_length) {
+  // for streaming (e.g. SSE stream sent to hystrix dashboard), we do not want
+  // chunk transfer encoding but we don't have a content-length so we pass "envoy only"
+  // header to avoid adding chunks
+  if (saw_content_length || headers.NoChunks()) {
     chunk_encoding_ = false;
   } else {
     if (processing_100_continue_) {